# src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Math Functions API
  version: 1.0.0
  description: API для работы с математическими функциями
servers:
  - url: http://localhost:8080/api
    description: Local server
security:
  - basicAuth: []
paths:
...
  /operations/add:
    post:
      summary: Сложение двух функций
      tags: [Operations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                function1Id:
                  type: integer
                function2Id:
                  type: integer
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunctionResponse'
  /operations/subtract:
    post:
      summary: Вычитание функций
      tags: [Operations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                function1Id:
                  type: integer
                function2Id:
                  type: integer
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunctionResponse'
  /operations/multiply:
    post:
      summary: Умножение функций
      tags: [Operations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                function1Id:
                  type: integer
                function2Id:
                  type: integer
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunctionResponse'
  /operations/divide:
    post:
      summary: Деление функций
      tags: [Operations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                function1Id:
                  type: integer
                function2Id:
                  type: integer
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunctionResponse'
  /operations/differentiate:
    post:
      summary: Продифференцировать функцию
      tags: [Operations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                functionId:
                  type: integer
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunctionResponse'
  /operations/integrate:
    post:
      summary: Вычислить интеграл функции
      tags: [Operations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                functionId:
                  type: integer
                threadsCount:
                  type: integer
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrateResponse'

  # ЭКСПОРТ/ИМПОРТ
...
components:
  schemas:
...
    FunctionResponse:
      type: object
      properties:
        functionId:
          type: integer
        functionName:
          type: string
        functionType:
          type: string
        ownerId:
          type: integer
        isPublic:
          type: boolean
        pointsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    IntegrateResponse:
      type: object
      properties:
        result:
          type: number
        computationTime:
          type: number
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Basic authentication with username and password